# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Ad Marketplace API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/user/name": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Update user name",
                "parameters": [
                    {
                        "description": "New name",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UpdateNameRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/user/profile": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get user profile",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ProfileResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/user/settings": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Update user settings",
                "parameters": [
                    {
                        "description": "Settings to update",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UpdateSettingsRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/user/wallet": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Link TON wallet",
                "parameters": [
                    {
                        "description": "Wallet address",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/LinkWalletRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "user"
                ],
                "summary": "Unlink TON wallet",
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "ErrorResponse": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "error_code": {
                    "type": "string"
                }
            }
        },
        "Language": {
            "type": "string",
            "enum": [
                "en",
                "ru"
            ],
            "x-enum-varnames": [
                "LanguageEN",
                "LanguageRU"
            ]
        },
...
        "LinkWalletRequest": {
            "type": "object",
            "required": [
                "address"
            ],
            "properties": {
                "address": {
                    "type": "string"
                }
            }
        },
...
        "PreferredMode": {
            "type": "string",
            "enum": [
                "publisher",
                "advertiser"
            ],
            "x-enum-varnames": [
                "PreferredModePublisher",
                "PreferredModeAdvertiser"
            ]
        },
        "ProfileResponse": {
            "type": "object",
            "properties": {
                "language": {
                    "$ref": "#/definitions/Language"
                },
                "name": {
                    "type": "string"
                },
                "onboarding_finished": {
                    "type": "boolean"
                },
                "preferred_mode": {
                    "$ref": "#/definitions/PreferredMode"
                },
                "receive_notifications": {
                    "type": "boolean"
                },
                "telegram_id": {
                    "type": "integer"
                },
                "theme": {
                    "$ref": "#/definitions/Theme"
                },
                "wallet_address": {
                    "type": "string"
                }
            }
        },
...
        "Theme": {
            "type": "string",
            "enum": [
                "light",
                "dark",
                "auto"
            ],
            "x-enum-varnames": [
                "ThemeLight",
                "ThemeDark",
                "ThemeAuto"
            ]
        },
...
        "UpdateNameRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "UpdateSettingsRequest": {
            "type": "object",
            "properties": {
                "language": {
                    "$ref": "#/definitions/Language"
                },
                "onboarding_finished": {
                    "type": "boolean"
                },
                "preferred_mode": {
                    "$ref": "#/definitions/PreferredMode"
                },
                "receive_notifications": {
                    "type": "boolean"
                },
                "theme": {
                    "$ref": "#/definitions/Theme"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}